ban phase in old draft is now a spearate component
